correct after check 2

+++ test-students-annotation/src/main/java/otus/student/kryukov/dz/check/Checker.java
@@ -4,6 +4,6 @@ import otus.student.kryukov.dz.domain.Question;
 
 public interface Checker {
 
-    public byte check(Question question, String answer);
+    public int check(Question question, String answer);
 
 }
